Kernel: add new class named `Kernel`
Normally a segment (module) is described by an instance of `Module`. However, for
launcher (or native VM) to load and transfer execution to smalltalk code. it has to
know locations to certain objects such as `nil`, `true`, `false` (and few others,
more may come in future, especially to support native VM).

To solve this, the kernel segment is described by a `Module` subclass: `Kernel` which
has a bunch of extra instance variables holding on these special objects (pretty
much like special objects array in OSVM).

Signed-off-by: Jan Vrany <jan.vrany@labware.com>